adding hooks and context
