+++ my-app/src/hooks/useProjects.jsx
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { ProjectContext } from "../contexts/ProjectContext";
+
+const useProjects = () => {
+  return useContext(ProjectContext);
+};
+
+export default useProjects;
